feat: refactor database schema organization and improve user service search functionality
