--- src/db/schema.ts
@@ -1,94 +1,0 @@
-import { sql } from 'drizzle-orm';
-import {
-  boolean,
-  foreignKey,
-  pgTable,
-  text,
-  timestamp,
-  unique,
-} from 'drizzle-orm/pg-core';
-
-export const session = pgTable(
-  'session',
-  {
-    id: text().primaryKey().notNull(),
-    expiresAt: timestamp({ mode: 'string' }).notNull(),
-    token: text().notNull(),
-    createdAt: timestamp({ mode: 'string' })
-      .default(sql`CURRENT_TIMESTAMP`)
-      .notNull(),
-    updatedAt: timestamp({ mode: 'string' }).notNull(),
-    ipAddress: text(),
-    userAgent: text(),
-    userId: text().notNull(),
-  },
-  (table) => [
-    foreignKey({
-      columns: [table.userId],
-      foreignColumns: [user.id],
-      name: 'session_userId_fkey',
-    }).onDelete('cascade'),
-    unique('session_token_key').on(table.token),
-  ],
-);
-
-export const account = pgTable(
-  'account',
-  {
-    id: text().primaryKey().notNull(),
-    accountId: text().notNull(),
-    providerId: text().notNull(),
-    userId: text().notNull(),
-    accessToken: text(),
-    refreshToken: text(),
-    idToken: text(),
-    accessTokenExpiresAt: timestamp({ mode: 'string' }),
-    refreshTokenExpiresAt: timestamp({ mode: 'string' }),
-    scope: text(),
-    password: text(),
-    createdAt: timestamp({ mode: 'string' })
-      .default(sql`CURRENT_TIMESTAMP`)
-      .notNull(),
-    updatedAt: timestamp({ mode: 'string' }).notNull(),
-  },
-  (table) => [
-    foreignKey({
-      columns: [table.userId],
-      foreignColumns: [user.id],
-      name: 'account_userId_fkey',
-    }).onDelete('cascade'),
-  ],
-);
-
-export const verification = pgTable('verification', {
-  id: text().primaryKey().notNull(),
-  identifier: text().notNull(),
-  value: text().notNull(),
-  expiresAt: timestamp({ mode: 'string' }).notNull(),
-  createdAt: timestamp({ mode: 'string' })
-    .default(sql`CURRENT_TIMESTAMP`)
-    .notNull(),
-  updatedAt: timestamp({ mode: 'string' })
-    .default(sql`CURRENT_TIMESTAMP`)
-    .notNull(),
-});
-
-export const user = pgTable(
-  'user',
-  {
-    id: text().primaryKey().notNull(),
-    name: text().notNull(),
-    email: text().notNull(),
-    emailVerified: boolean().notNull(),
-    image: text(),
-    createdAt: timestamp({ mode: 'string' })
-      .default(sql`CURRENT_TIMESTAMP`)
-      .notNull(),
-    updatedAt: timestamp({ mode: 'string' })
-      .default(sql`CURRENT_TIMESTAMP`)
-      .notNull(),
-  },
-  (table) => [unique('user_email_key').on(table.email)],
-);
-
-export const schema = { session, account, verification, user };
