+++ src/db/schemas/index.ts
@@ -1,0 +1,7 @@
+import { user } from './user';
+import { session } from './session';
+import { account } from './account';
+import { verification } from './verification';
+
+export { user, session, account, verification };
+export const schema = { user, session, account, verification };
