+++ src/lib/auth/config.ts
@@ -2,7 +2,7 @@ import { betterAuth } from 'better-auth';
 import { drizzleAdapter } from 'better-auth/adapters/drizzle';
 import { nextCookies } from 'better-auth/next-js';
 import { db } from '~/db/drizzle';
-import { schema } from '~/db/schema';
+import { schema } from '~/db/schemas';
 
 export const auth = betterAuth({
   database: drizzleAdapter(db, {
